pro_final/requirements.txt
appdirs==1.4.3
boltons==16.1.1
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.7
django-taggit==0.17.1
django-toolbelt==0.0.1
fixmydjango==0.2
flake8==3.3.0
gunicorn==19.7.1
mccabe==0.6.1
mock==2.0.0
model-mommy==1.3.2
olefile==0.44
packaging==16.8
pbr==3.0.0
Pillow==4.1.1
psycopg2==2.7.1
pycodestyle==2.3.1
pyflakes==1.5.0
pyparsing==2.2.0
requests==2.9.1
six==1.10.0
static3==0.7.0
termcolor==1.1.0
